+++ Packages/Syntax/Sources/SyntaxParsers/Queries/Python/outline.scm
@@ -12,6 +12,10 @@
 (class_definition
   name: (identifier) @outline.container)
 
+; Values
+(assignment
+  left: (identifier) @outline.value)
+
 ; Functions
 (function_definition
   name: (identifier) @outline.function